better HTTP error handling
