+++ lib/bulk-proxy-endpoint.js
@@ -1,5 +1,5 @@
 const {EventEmitter} = require("events");
-const Endpoint = require("./endpoint");
+const {post} = require("./http");
 const fib = require("./fib");
 const sleep = require("./sleep");
 const {assign, freeze} = Object;
@@ -32,8 +32,6 @@ class BulkProxyEndpoint extends EventEmitter {
             retries: undefined,
             slowThreshold: undefined
         };
-
-        this[$private].endpoint = Endpoint(this.url);
     }
 
     /**
@@ -313,7 +311,7 @@ function start(endpoint) {
 
             try {
                 timer = Date.now();
-                result = await endpoint[$private].endpoint(chunks);
+                result = await post(endpoint.url, chunks);
                 endpoint[$private].size -= length;
                 endpoint[$private].failed = undefined;
                 endpoint.emit("inserted", chunks);
